CR: categories,applying to jobs, job search by preferences.

+++ src/main/java/com/chicmic/eNaukri/model/Preference.java
@@ -14,8 +14,8 @@ public class Preference {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String location;
-    private Float salary;
-    private Float yoe;
+    private Integer salary;
+    private Integer yoe;
     private String remoteHybridOnsite;
     @OneToOne
     private UserProfile userPreferences;
